requirements.txt
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configs==3.0.3
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
requests==2.21.0
urllib3==1.24.2
Werkzeug==0.15.2
python-dateutil==2.8.0